requirements.txt
discord.py>=2.1.0
ijson>=3.3.0
pygge>=2.1.0
psycopg2-binary>=2.9.1